Improve typing and remove nested recursive import
Change-type: patch
